refactor: 자기소개는 구현
